will, i am sheikh sphere

+++ ray.h
@@ -8,8 +8,8 @@ typedef struct {
     vec3 dir;
 } ray;
 
-static inline point3 ray_at(ray r, double t) {
-    vec3 scale = (vec3_scale((r.dir), t));
+static inline point3 ray_at(ray r, float t) {
+    vec3 scale = vec3_scale((r.dir), t);
     return vec3_add(r.orig, scale);
 }
